+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- `OperationType` type
+
+- `ExecutionStagesTableData` type
+- 
+- `TrackingOperationResult` type
+
+- `StageName` enum
+
+- `getOperationType` in `OperationTracker` retrieves the `OperationType` for `operationId`
+
+
+### Changed
+
+- The stage names have been changed
+
+- Changed namings in enums 
+
+- `OperationType` added in the `ExecutionStages`
+
+- `ExecutionStages` structure
+
+- Added return value in method `sendShardTransaction` in `TonConnectSender`
+
+- Added `forceSend` option in method `sendCrossChainTransaction` in `TacSdk`
+
+- startTracking has been improved. Added optional parameters and return values
+
+### Removed
+
+- Deleted `isBridgeOperation`(now it can be determined with `getOperationType`)
+
 ## [0.5.0] - 2025-03-03
 
 ### Changed